# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VeriFactu SaaS API - Sistema de Facturación AEAT
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/taxpayers:
    get:
      operationId: listTaxpayers
      tags: [taxpayers]
      summary: List taxpayer profiles for organization
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Taxpayers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxpayerProfile"
    post:
      operationId: createTaxpayer
      tags: [taxpayers]
      summary: Create taxpayer profile
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxpayerBody"
      responses:
        "201":
          description: Taxpayer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxpayerProfile"
  /taxpayers/{id}:
    get:
      operationId: getTaxpayer
      tags: [taxpayers]
      summary: Get taxpayer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Taxpayer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxpayerProfile"
    patch:
      operationId: updateTaxpayer
      tags: [taxpayers]
      summary: Update taxpayer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaxpayerBody"
      responses:
        "200":
          description: Taxpayer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxpayerProfile"

  # Clients
...
  /taxpayers/{taxpayerId}/verifactu/records:
    get:
      operationId: listVeriFactuRecords
      tags: [verifactu]
      summary: List VeriFactu records
      parameters:
        - name: taxpayerId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: VeriFactu records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VeriFactuRecordListResponse"
  /verifactu/records/{id}:
    get:
      operationId: getVeriFactuRecord
      tags: [verifactu]
      summary: Get VeriFactu record detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: VeriFactu record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VeriFactuRecord"
  /verifactu/records/{id}/submit:
    post:
      operationId: submitVeriFactuRecord
      tags: [verifactu]
      summary: Submit VeriFactu record to AEAT
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VeriFactuRecord"
  /verifactu/records/{id}/retry:
    post:
      operationId: retryVeriFactuRecord
      tags: [verifactu]
      summary: Retry failed VeriFactu submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VeriFactuRecord"

  # Documents / OCR
...
components:
  schemas:
...
    TaxpayerProfile:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        tradeName:
          type: ["string", "null"]
        nif:
          type: string
        nifType:
          type: string
          enum: [NIF, NIE, CIF, PASSPORT, OTHER]
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
        country:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        defaultVatRate:
          type: number
        sifInstallationNumber:
          type: ["string", "null"]
        aeatEnvironment:
          type: string
          enum: [sandbox, production]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - name
        - nif
        - nifType
        - address
        - city
        - postalCode
        - province
        - country
        - defaultVatRate
        - aeatEnvironment
        - isActive
        - createdAt
        - updatedAt
    CreateTaxpayerBody:
      type: object
      properties:
        name:
          type: string
        tradeName:
          type: ["string", "null"]
        nif:
          type: string
        nifType:
          type: string
          enum: [NIF, NIE, CIF, PASSPORT, OTHER]
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
        country:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        defaultVatRate:
          type: number
      required:
        - name
        - nif
        - nifType
        - address
        - city
        - postalCode
        - province
        - country
    UpdateTaxpayerBody:
      type: object
      properties:
        name:
          type: string
        tradeName:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        defaultVatRate:
          type: number
        aeatEnvironment:
          type: string
          enum: [sandbox, production]

    # Client
...
    VeriFactuRecord:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        taxpayerId:
          type: integer
        recordType:
          type: string
          enum: [ALTA, ANULACION, SUBSANACION]
        status:
          type: string
          enum: [PENDING, SUBMITTED, ACCEPTED, ACCEPTED_WITH_ERRORS, REJECTED, ERROR]
        hash:
          type: ["string", "null"]
        previousHash:
          type: ["string", "null"]
        qrUrl:
          type: ["string", "null"]
        xmlPayload:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        aeatResponse:
          type: ["string", "null"]
        aeatCsv:
          type: ["string", "null"]
        aeatErrorCode:
          type: ["string", "null"]
        aeatErrorMessage:
          type: ["string", "null"]
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceId
        - taxpayerId
        - recordType
        - status
        - retryCount
        - createdAt
        - updatedAt
    VeriFactuRecordListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VeriFactuRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - pages

    # Documents
...