# lib/api-zod/src/generated/types/createTaxpayerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeriFactu SaaS API - Sistema de Facturación AEAT
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaxpayerBodyNifType } from "./createTaxpayerBodyNifType";

export interface CreateTaxpayerBody {
  name: string;
  /** @nullable */
  tradeName?: string | null;
  nif: string;
  nifType: CreateTaxpayerBodyNifType;
  address: string;
  city: string;
  postalCode: string;
  province: string;
  country: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  defaultVatRate?: number;
}